+++ src/bl64-lib.bash
@@ -47,9 +47,8 @@ function _bl64_lib_script_get_path() {
 
 function _bl64_lib_script_get_name() {
   local base=''
-
-  if [[ -n "$BL64_SCRIPT_PATH" && "$BL64_SCRIPT_PATH" != '/' ]]; then
-    base="${BL64_SCRIPT_PATH##*/}"
+  if [[ -n "$BASH_ARGV0" && "$BASH_ARGV0" != '/' ]]; then
+    base="${BASH_ARGV0##*/}"
   fi
   if [[ -z "$base" || "$base" == */* ]]; then
     base='noname'